# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TicketflowAI API",
        "version": "0.5"
    },
    "paths": {
...
        "/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get a list of companies",
                "operationId": "8af1177352a157e9228c00f023616d87",
                "responses": {
                    "200": {
                        "description": "List of companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a new company",
                "operationId": "dd2ae66c40ea62cd8f20d49bc08e4b21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "idNumber",
                                    "contactEmail",
                                    "phone",
                                    "state",
                                    "city",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "MindSoft Inc."
                                    },
                                    "idNumber": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "contactEmail": {
                                        "type": "string",
                                        "example": "contact@mindsoft.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1-555-1234"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "California"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "San Francisco"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "1234 Market Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get details of a specific company",
                "operationId": "662432d4be675d549b84ca1bf7e56b8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the company",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update a company",
                "operationId": "00dbc7963c0f6ac45bd2b2595fb1f88e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the company",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "idNumber": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "contactEmail": {
                                        "type": "string",
                                        "example": "updated@mindsoft.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1-555-6789"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Brooklyn"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "5678 Main Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "operationId": "00a9e816ead34effffa2a6378e1cde4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the company",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Company deleted successfully"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompanyResource": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the company",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the company",
                        "type": "string",
                        "example": "MindSoft Inc."
                    },
                    "idNumber": {
                        "description": "Identification number of the company",
                        "type": "string",
                        "example": "123456789"
                    },
                    "contactEmail": {
                        "description": "Contact email of the company",
                        "type": "string",
                        "example": "contact@mindsoft.com"
                    },
                    "phone": {
                        "description": "Contact phone number of the company",
                        "type": "string",
                        "example": "+1-555-1234"
                    },
                    "state": {
                        "description": "State where the company is located",
                        "type": "string",
                        "example": "California"
                    },
                    "city": {
                        "description": "City where the company is located",
                        "type": "string",
                        "example": "San Francisco"
                    },
                    "address": {
                        "description": "Address of the company",
                        "type": "string",
                        "example": "1234 Market Street"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}